Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/css/css-fonts/parsing/font-valid-expected.txt


PASS caption should be a supported system font.
PASS icon should be a supported system font.
PASS menu should be a supported system font.
PASS message-box should be a supported system font.
PASS small-caption should be a supported system font.
PASS status-bar should be a supported system font.
PASS e.style['font'] = "xx-small serif" should set the property value
FAIL e.style['font'] = "normal medium/normal sans-serif" should set the property value assert_equals: serialization should be canonical expected "medium sans-serif" but got "normal medium/normal sans-serif"
FAIL e.style['font'] = "normal normal xx-large/1.2 cursive" should set the property value assert_equals: serialization should be canonical expected "xx-large / 1.2 cursive" but got "normal normal xx-large/1.2 cursive"
FAIL e.style['font'] = "normal normal normal larger/calc(120% + 1.2em) fantasy" should set the property value assert_equals: serialization should be canonical expected "larger / calc(120% + 1.2em) fantasy" but got "normal normal normal larger/calc(120% + 1.2em) fantasy"
FAIL e.style['font'] = "normal normal normal normal smaller monospace" should set the property value assert_equals: serialization should be canonical expected "smaller monospace" but got "normal normal normal normal smaller monospace"
FAIL e.style['font'] = "normal normal normal italic 10px/normal Menu" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal normal normal small-caps 20%/1.2 \"FB Armada\"" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal normal normal bold calc(30% - 40px)/calc(120% + 1.2em) serif" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal normal normal ultra-condensed xx-small sans-serif" should set the property value assert_equals: serialization should be canonical expected "ultra-condensed xx-small sans-serif" but got "normal normal normal ultra-condensed xx-small sans-serif"
FAIL e.style['font'] = "normal normal oblique medium/normal cursive" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal normal italic normal xx-large/1.2 fantasy" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal normal oblique small-caps larger/calc(120% + 1.2em) monospace" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal normal italic bolder smaller Menu" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal normal oblique extra-condensed 10px/normal \"FB Armada\"" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal normal small-caps 20%/1.2 serif" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal normal small-caps normal calc(30% - 40px)/calc(120% + 1.2em) sans-serif" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal normal small-caps italic xx-small cursive" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal normal small-caps lighter medium/normal fantasy" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal normal small-caps condensed xx-large/1.2 monospace" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal normal 100 larger/calc(120% + 1.2em) Menu" should set the property value assert_equals: serialization should be canonical expected "100 larger / calc(120% + 1.2em) Menu" but got "normal normal 100 larger/calc(120% + 1.2em) Menu"
FAIL e.style['font'] = "normal normal 900 normal smaller \"FB Armada\"" should set the property value assert_equals: serialization should be canonical expected "900 smaller \"FB Armada\"" but got "normal normal 900 normal smaller \"FB Armada\""
FAIL e.style['font'] = "normal normal bold oblique 10px/normal serif" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal normal bolder small-caps 20%/1.2 sans-serif" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal normal lighter semi-condensed calc(30% - 40px)/calc(120% + 1.2em) cursive" should set the property value assert_equals: serialization should be canonical expected "lighter semi-condensed calc(30% - 40px) / calc(120% + 1.2em) cursive" but got "normal normal lighter semi-condensed calc(30% - 40px)/calc(120% + 1.2em) cursive"
FAIL e.style['font'] = "normal normal semi-expanded xx-small fantasy" should set the property value assert_equals: serialization should be canonical expected "semi-expanded xx-small fantasy" but got "normal normal semi-expanded xx-small fantasy"
FAIL e.style['font'] = "normal normal expanded normal medium/normal monospace" should set the property value assert_equals: serialization should be canonical expected "expanded medium monospace" but got "normal normal normal expanded medium/normal monospace"
FAIL e.style['font'] = "normal normal extra-expanded italic xx-large/1.2 Menu" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal normal ultra-expanded small-caps larger/calc(120% + 1.2em) \"FB Armada\"" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal normal ultra-condensed 100 smaller serif" should set the property value assert_equals: serialization should be canonical expected "100 ultra-condensed smaller serif" but got "normal normal 100 ultra-condensed smaller serif"
FAIL e.style['font'] = "normal oblique 10px/normal sans-serif" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal italic normal 20%/1.2 cursive" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal oblique normal normal calc(30% - 40px)/calc(120% + 1.2em) fantasy" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal italic normal small-caps xx-small monospace" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal oblique normal 900 medium/normal Menu" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal italic normal extra-condensed xx-large/1.2 \"FB Armada\"" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal oblique small-caps larger/calc(120% + 1.2em) serif" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal italic small-caps normal smaller sans-serif" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal oblique small-caps bold 10px/normal cursive" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal italic small-caps condensed 20%/1.2 fantasy" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal oblique bolder calc(30% - 40px)/calc(120% + 1.2em) monospace" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal italic lighter normal xx-small Menu" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal oblique 100 small-caps medium/normal \"FB Armada\"" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal italic 900 semi-condensed xx-large/1.2 serif" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal oblique semi-expanded larger/calc(120% + 1.2em) sans-serif" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal italic expanded normal smaller cursive" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal oblique extra-expanded small-caps 10px/normal fantasy" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal italic ultra-expanded bold 20%/1.2 monospace" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal small-caps calc(30% - 40px)/calc(120% + 1.2em) Menu" should set the property value assert_equals: serialization should be canonical expected "small-caps calc(30% - 40px) / calc(120% + 1.2em) Menu" but got "normal small-caps calc(30% - 40px)/calc(120% + 1.2em) Menu"
FAIL e.style['font'] = "normal small-caps normal xx-small \"FB Armada\"" should set the property value assert_equals: serialization should be canonical expected "small-caps xx-small \"FB Armada\"" but got "normal small-caps normal xx-small \"FB Armada\""
FAIL e.style['font'] = "normal small-caps normal normal medium/normal serif" should set the property value assert_equals: serialization should be canonical expected "small-caps medium serif" but got "normal small-caps normal normal medium/normal serif"
FAIL e.style['font'] = "normal small-caps normal oblique xx-large/1.2 sans-serif" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal small-caps normal bolder larger/calc(120% + 1.2em) cursive" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal small-caps normal ultra-condensed smaller fantasy" should set the property value assert_equals: serialization should be canonical expected "small-caps ultra-condensed smaller fantasy" but got "normal small-caps normal ultra-condensed smaller fantasy"
FAIL e.style['font'] = "normal small-caps italic 10px/normal monospace" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal small-caps oblique normal 20%/1.2 Menu" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal small-caps italic lighter calc(30% - 40px)/calc(120% + 1.2em) \"FB Armada\"" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal small-caps oblique extra-condensed xx-small serif" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal small-caps 100 medium/normal sans-serif" should set the property value assert_equals: serialization should be canonical expected "small-caps 100 medium sans-serif" but got "normal small-caps 100 medium/normal sans-serif"
FAIL e.style['font'] = "normal small-caps 900 normal xx-large/1.2 cursive" should set the property value assert_equals: serialization should be canonical expected "small-caps 900 xx-large / 1.2 cursive" but got "normal small-caps 900 normal xx-large/1.2 cursive"
FAIL e.style['font'] = "normal small-caps bold italic larger/calc(120% + 1.2em) fantasy" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal small-caps bolder condensed smaller monospace" should set the property value assert_equals: serialization should be canonical expected "small-caps bolder condensed smaller monospace" but got "normal small-caps bolder condensed smaller monospace"
FAIL e.style['font'] = "normal small-caps semi-condensed 10px/normal Menu" should set the property value assert_equals: serialization should be canonical expected "small-caps semi-condensed 10px Menu" but got "normal small-caps semi-condensed 10px/normal Menu"
FAIL e.style['font'] = "normal small-caps semi-expanded normal 20%/1.2 \"FB Armada\"" should set the property value assert_equals: serialization should be canonical expected "small-caps semi-expanded 20% / 1.2 \"FB Armada\"" but got "normal small-caps normal semi-expanded 20%/1.2 \"FB Armada\""
FAIL e.style['font'] = "normal small-caps expanded oblique calc(30% - 40px)/calc(120% + 1.2em) serif" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal small-caps extra-expanded lighter xx-small sans-serif" should set the property value assert_equals: serialization should be canonical expected "small-caps lighter extra-expanded xx-small sans-serif" but got "normal small-caps lighter extra-expanded xx-small sans-serif"
FAIL e.style['font'] = "normal 100 medium/normal cursive" should set the property value assert_equals: serialization should be canonical expected "100 medium cursive" but got "normal 100 medium/normal cursive"
FAIL e.style['font'] = "normal 900 normal xx-large/1.2 fantasy" should set the property value assert_equals: serialization should be canonical expected "900 xx-large / 1.2 fantasy" but got "normal normal 900 xx-large/1.2 fantasy"
FAIL e.style['font'] = "normal bold normal normal larger/calc(120% + 1.2em) monospace" should set the property value assert_equals: serialization should be canonical expected "bold larger / calc(120% + 1.2em) monospace" but got "normal normal bold normal larger/calc(120% + 1.2em) monospace"
FAIL e.style['font'] = "normal bolder normal italic smaller Menu" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal lighter normal small-caps 10px/normal \"FB Armada\"" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal 100 normal ultra-expanded 20%/1.2 serif" should set the property value assert_equals: serialization should be canonical expected "100 ultra-expanded 20% / 1.2 serif" but got "normal normal 100 ultra-expanded 20%/1.2 serif"
FAIL e.style['font'] = "normal 900 oblique calc(30% - 40px)/calc(120% + 1.2em) sans-serif" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal bold italic normal xx-small cursive" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal bolder oblique small-caps medium/normal fantasy" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal lighter italic ultra-condensed xx-large/1.2 monospace" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal 100 small-caps larger/calc(120% + 1.2em) Menu" should set the property value assert_equals: serialization should be canonical expected "small-caps 100 larger / calc(120% + 1.2em) Menu" but got "normal small-caps 100 larger/calc(120% + 1.2em) Menu"
FAIL e.style['font'] = "normal 900 small-caps normal smaller \"FB Armada\"" should set the property value assert_equals: serialization should be canonical expected "small-caps 900 smaller \"FB Armada\"" but got "normal small-caps 900 normal smaller \"FB Armada\""
FAIL e.style['font'] = "normal bold small-caps oblique 10px/normal serif" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal bolder small-caps extra-condensed 20%/1.2 sans-serif" should set the property value assert_equals: serialization should be canonical expected "small-caps bolder extra-condensed 20% / 1.2 sans-serif" but got "normal small-caps bolder extra-condensed 20%/1.2 sans-serif"
FAIL e.style['font'] = "normal lighter condensed calc(30% - 40px)/calc(120% + 1.2em) cursive" should set the property value assert_equals: serialization should be canonical expected "lighter condensed calc(30% - 40px) / calc(120% + 1.2em) cursive" but got "normal lighter condensed calc(30% - 40px)/calc(120% + 1.2em) cursive"
FAIL e.style['font'] = "normal 100 semi-condensed normal xx-small fantasy" should set the property value assert_equals: serialization should be canonical expected "100 semi-condensed xx-small fantasy" but got "normal normal 100 semi-condensed xx-small fantasy"
FAIL e.style['font'] = "normal 900 semi-expanded italic medium/normal monospace" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal bold expanded small-caps xx-large/1.2 Menu" should set the property value assert_equals: serialization should be canonical expected "small-caps bold expanded xx-large / 1.2 Menu" but got "normal small-caps bold expanded xx-large/1.2 Menu"
FAIL e.style['font'] = "normal extra-expanded larger/calc(120% + 1.2em) \"FB Armada\"" should set the property value assert_equals: serialization should be canonical expected "extra-expanded larger / calc(120% + 1.2em) \"FB Armada\"" but got "normal extra-expanded larger/calc(120% + 1.2em) \"FB Armada\""
FAIL e.style['font'] = "normal ultra-expanded normal smaller serif" should set the property value assert_equals: serialization should be canonical expected "ultra-expanded smaller serif" but got "normal normal ultra-expanded smaller serif"
FAIL e.style['font'] = "normal ultra-condensed normal normal 10px/normal sans-serif" should set the property value assert_equals: serialization should be canonical expected "ultra-condensed 10px sans-serif" but got "normal normal normal ultra-condensed 10px/normal sans-serif"
FAIL e.style['font'] = "normal extra-condensed normal oblique 20%/1.2 cursive" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal condensed normal small-caps calc(30% - 40px)/calc(120% + 1.2em) fantasy" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal semi-condensed normal bolder xx-small monospace" should set the property value assert_equals: serialization should be canonical expected "bolder semi-condensed xx-small monospace" but got "normal normal bolder semi-condensed xx-small monospace"
FAIL e.style['font'] = "normal semi-expanded italic medium/normal Menu" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal expanded oblique normal xx-large/1.2 \"FB Armada\"" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal extra-expanded italic small-caps larger/calc(120% + 1.2em) serif" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal ultra-expanded oblique lighter smaller sans-serif" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal ultra-condensed small-caps 10px/normal cursive" should set the property value assert_equals: serialization should be canonical expected "small-caps ultra-condensed 10px cursive" but got "normal small-caps ultra-condensed 10px/normal cursive"
FAIL e.style['font'] = "normal extra-condensed small-caps normal 20%/1.2 fantasy" should set the property value assert_equals: serialization should be canonical expected "small-caps extra-condensed 20% / 1.2 fantasy" but got "normal small-caps normal extra-condensed 20%/1.2 fantasy"
FAIL e.style['font'] = "normal condensed small-caps italic calc(30% - 40px)/calc(120% + 1.2em) monospace" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal semi-condensed small-caps 100 xx-small Menu" should set the property value assert_equals: serialization should be canonical expected "small-caps 100 semi-condensed xx-small Menu" but got "normal small-caps 100 semi-condensed xx-small Menu"
FAIL e.style['font'] = "normal semi-expanded 900 medium/normal \"FB Armada\"" should set the property value assert_equals: serialization should be canonical expected "900 semi-expanded medium \"FB Armada\"" but got "normal 900 semi-expanded medium/normal \"FB Armada\""
FAIL e.style['font'] = "normal expanded bold normal xx-large/1.2 serif" should set the property value assert_equals: serialization should be canonical expected "bold expanded xx-large / 1.2 serif" but got "normal normal bold expanded xx-large/1.2 serif"
FAIL e.style['font'] = "normal extra-expanded bolder oblique larger/calc(120% + 1.2em) sans-serif" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "normal ultra-expanded lighter small-caps smaller cursive" should set the property value assert_equals: serialization should be canonical expected "small-caps lighter ultra-expanded smaller cursive" but got "normal small-caps lighter ultra-expanded smaller cursive"
FAIL e.style['font'] = "italic 10px/normal fantasy" should set the property value assert_equals: serialization should be canonical expected "italic 10px fantasy" but got "italic 10px/normal fantasy"
FAIL e.style['font'] = "oblique normal 20%/1.2 monospace" should set the property value assert_equals: serialization should be canonical expected "oblique 20% / 1.2 monospace" but got "oblique normal 20%/1.2 monospace"
FAIL e.style['font'] = "italic normal normal calc(30% - 40px)/calc(120% + 1.2em) Menu" should set the property value assert_equals: serialization should be canonical expected "italic calc(30% - 40px) / calc(120% + 1.2em) Menu" but got "italic normal normal calc(30% - 40px)/calc(120% + 1.2em) Menu"
FAIL e.style['font'] = "oblique normal normal normal xx-small \"FB Armada\"" should set the property value assert_equals: serialization should be canonical expected "oblique xx-small \"FB Armada\"" but got "oblique normal normal normal xx-small \"FB Armada\""
FAIL e.style['font'] = "italic normal normal small-caps medium/normal serif" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "oblique normal normal 100 xx-large/1.2 sans-serif" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "italic normal normal ultra-condensed larger/calc(120% + 1.2em) cursive" should set the property value assert_equals: serialization should be canonical expected "italic ultra-condensed larger / calc(120% + 1.2em) cursive" but got "italic normal normal ultra-condensed larger/calc(120% + 1.2em) cursive"
FAIL e.style['font'] = "oblique normal small-caps smaller fantasy" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "italic normal small-caps normal 10px/normal monospace" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "oblique normal small-caps 900 20%/1.2 Menu" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "italic normal small-caps extra-condensed calc(30% - 40px)/calc(120% + 1.2em) \"FB Armada\"" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "oblique normal bold xx-small serif" should set the property value assert_equals: serialization should be canonical expected "oblique bold xx-small serif" but got "oblique normal bold xx-small serif"
FAIL e.style['font'] = "italic normal bolder normal medium/normal sans-serif" should set the property value assert_equals: serialization should be canonical expected "italic bolder medium sans-serif" but got "italic normal bolder normal medium/normal sans-serif"
FAIL e.style['font'] = "oblique normal lighter small-caps xx-large/1.2 cursive" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "italic normal 100 condensed larger/calc(120% + 1.2em) fantasy" should set the property value assert_equals: serialization should be canonical expected "italic 100 condensed larger / calc(120% + 1.2em) fantasy" but got "italic normal 100 condensed larger/calc(120% + 1.2em) fantasy"
FAIL e.style['font'] = "oblique normal semi-condensed smaller monospace" should set the property value assert_equals: serialization should be canonical expected "oblique semi-condensed smaller monospace" but got "oblique normal semi-condensed smaller monospace"
FAIL e.style['font'] = "italic normal semi-expanded normal 10px/normal Menu" should set the property value assert_equals: serialization should be canonical expected "italic semi-expanded 10px Menu" but got "italic normal normal semi-expanded 10px/normal Menu"
FAIL e.style['font'] = "oblique normal expanded small-caps 20%/1.2 \"FB Armada\"" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "italic normal extra-expanded 900 calc(30% - 40px)/calc(120% + 1.2em) serif" should set the property value assert_equals: serialization should be canonical expected "italic 900 extra-expanded calc(30% - 40px) / calc(120% + 1.2em) serif" but got "italic normal 900 extra-expanded calc(30% - 40px)/calc(120% + 1.2em) serif"
PASS e.style['font'] = "oblique small-caps xx-small sans-serif" should set the property value
FAIL e.style['font'] = "italic small-caps normal medium/normal cursive" should set the property value assert_equals: serialization should be canonical expected "italic small-caps medium cursive" but got "italic small-caps normal medium/normal cursive"
FAIL e.style['font'] = "oblique small-caps normal normal xx-large/1.2 fantasy" should set the property value assert_equals: serialization should be canonical expected "oblique small-caps xx-large / 1.2 fantasy" but got "oblique small-caps normal normal xx-large/1.2 fantasy"
FAIL e.style['font'] = "italic small-caps normal bold larger/calc(120% + 1.2em) monospace" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "oblique small-caps normal ultra-expanded smaller Menu" should set the property value assert_equals: serialization should be canonical expected "oblique small-caps ultra-expanded smaller Menu" but got "oblique small-caps normal ultra-expanded smaller Menu"
FAIL e.style['font'] = "italic small-caps bolder 10px/normal \"FB Armada\"" should set the property value assert_equals: serialization should be canonical expected "italic small-caps bolder 10px \"FB Armada\"" but got "italic small-caps bolder 10px/normal \"FB Armada\""
FAIL e.style['font'] = "oblique small-caps lighter normal 20%/1.2 serif" should set the property value assert_equals: serialization should be canonical expected "oblique small-caps lighter 20% / 1.2 serif" but got "oblique small-caps lighter normal 20%/1.2 serif"
FAIL e.style['font'] = "italic small-caps 100 ultra-condensed calc(30% - 40px)/calc(120% + 1.2em) sans-serif" should set the property value assert_equals: serialization should be canonical expected "italic small-caps 100 ultra-condensed calc(30% - 40px) / calc(120% + 1.2em) sans-serif" but got "italic small-caps 100 ultra-condensed calc(30% - 40px)/calc(120% + 1.2em) sans-serif"
PASS e.style['font'] = "oblique small-caps extra-condensed xx-small cursive" should set the property value
FAIL e.style['font'] = "italic small-caps condensed normal medium/normal fantasy" should set the property value assert_equals: serialization should be canonical expected "italic small-caps condensed medium fantasy" but got "italic small-caps normal condensed medium/normal fantasy"
FAIL e.style['font'] = "oblique small-caps semi-condensed 900 xx-large/1.2 monospace" should set the property value assert_equals: serialization should be canonical expected "oblique small-caps 900 semi-condensed xx-large / 1.2 monospace" but got "oblique small-caps 900 semi-condensed xx-large/1.2 monospace"
FAIL e.style['font'] = "italic bold larger/calc(120% + 1.2em) Menu" should set the property value assert_equals: serialization should be canonical expected "italic bold larger / calc(120% + 1.2em) Menu" but got "italic bold larger/calc(120% + 1.2em) Menu"
FAIL e.style['font'] = "oblique bolder normal smaller \"FB Armada\"" should set the property value assert_equals: serialization should be canonical expected "oblique bolder smaller \"FB Armada\"" but got "oblique normal bolder smaller \"FB Armada\""
FAIL e.style['font'] = "italic lighter normal normal 10px/normal serif" should set the property value assert_equals: serialization should be canonical expected "italic lighter 10px serif" but got "italic normal lighter normal 10px/normal serif"
FAIL e.style['font'] = "oblique 100 normal small-caps 20%/1.2 sans-serif" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "italic 900 normal semi-expanded calc(30% - 40px)/calc(120% + 1.2em) cursive" should set the property value assert_equals: serialization should be canonical expected "italic 900 semi-expanded calc(30% - 40px) / calc(120% + 1.2em) cursive" but got "italic normal 900 semi-expanded calc(30% - 40px)/calc(120% + 1.2em) cursive"
PASS e.style['font'] = "oblique bold small-caps xx-small fantasy" should set the property value
FAIL e.style['font'] = "italic bolder small-caps normal medium/normal monospace" should set the property value assert_equals: serialization should be canonical expected "italic small-caps bolder medium monospace" but got "italic small-caps bolder normal medium/normal monospace"
FAIL e.style['font'] = "oblique lighter small-caps expanded xx-large/1.2 Menu" should set the property value assert_equals: serialization should be canonical expected "oblique small-caps lighter expanded xx-large / 1.2 Menu" but got "oblique small-caps lighter expanded xx-large/1.2 Menu"
FAIL e.style['font'] = "italic 100 extra-expanded larger/calc(120% + 1.2em) \"FB Armada\"" should set the property value assert_equals: serialization should be canonical expected "italic 100 extra-expanded larger / calc(120% + 1.2em) \"FB Armada\"" but got "italic 100 extra-expanded larger/calc(120% + 1.2em) \"FB Armada\""
FAIL e.style['font'] = "oblique 900 ultra-expanded normal smaller serif" should set the property value assert_equals: serialization should be canonical expected "oblique 900 ultra-expanded smaller serif" but got "oblique normal 900 ultra-expanded smaller serif"
FAIL e.style['font'] = "italic bold ultra-condensed small-caps 10px/normal sans-serif" should set the property value assert_equals: serialization should be canonical expected "italic small-caps bold ultra-condensed 10px sans-serif" but got "italic small-caps bold ultra-condensed 10px/normal sans-serif"
FAIL e.style['font'] = "oblique extra-condensed 20%/1.2 cursive" should set the property value assert_equals: serialization should be canonical expected "oblique extra-condensed 20% / 1.2 cursive" but got "oblique extra-condensed 20%/1.2 cursive"
FAIL e.style['font'] = "italic condensed normal calc(30% - 40px)/calc(120% + 1.2em) fantasy" should set the property value assert_equals: serialization should be canonical expected "italic condensed calc(30% - 40px) / calc(120% + 1.2em) fantasy" but got "italic normal condensed calc(30% - 40px)/calc(120% + 1.2em) fantasy"
FAIL e.style['font'] = "oblique semi-condensed normal normal xx-small monospace" should set the property value assert_equals: serialization should be canonical expected "oblique semi-condensed xx-small monospace" but got "oblique normal normal semi-condensed xx-small monospace"
FAIL e.style['font'] = "italic semi-expanded normal small-caps medium/normal Menu" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "oblique expanded normal bolder xx-large/1.2 \"FB Armada\"" should set the property value assert_equals: serialization should be canonical expected "oblique bolder expanded xx-large / 1.2 \"FB Armada\"" but got "oblique normal bolder expanded xx-large/1.2 \"FB Armada\""
FAIL e.style['font'] = "italic extra-expanded small-caps larger/calc(120% + 1.2em) serif" should set the property value assert_equals: serialization should be canonical expected "italic small-caps extra-expanded larger / calc(120% + 1.2em) serif" but got "italic small-caps extra-expanded larger/calc(120% + 1.2em) serif"
FAIL e.style['font'] = "oblique ultra-expanded small-caps normal smaller sans-serif" should set the property value assert_equals: serialization should be canonical expected "oblique small-caps ultra-expanded smaller sans-serif" but got "oblique small-caps normal ultra-expanded smaller sans-serif"
FAIL e.style['font'] = "italic ultra-condensed small-caps lighter 10px/normal cursive" should set the property value assert_equals: serialization should be canonical expected "italic small-caps lighter ultra-condensed 10px cursive" but got "italic small-caps lighter ultra-condensed 10px/normal cursive"
FAIL e.style['font'] = "oblique extra-condensed 100 20%/1.2 fantasy" should set the property value assert_equals: serialization should be canonical expected "oblique 100 extra-condensed 20% / 1.2 fantasy" but got "oblique 100 extra-condensed 20%/1.2 fantasy"
FAIL e.style['font'] = "italic condensed 900 normal calc(30% - 40px)/calc(120% + 1.2em) monospace" should set the property value assert_equals: serialization should be canonical expected "italic 900 condensed calc(30% - 40px) / calc(120% + 1.2em) monospace" but got "italic normal 900 condensed calc(30% - 40px)/calc(120% + 1.2em) monospace"
PASS e.style['font'] = "oblique semi-condensed bold small-caps xx-small Menu" should set the property value
FAIL e.style['font'] = "small-caps medium/normal \"FB Armada\"" should set the property value assert_equals: serialization should be canonical expected "small-caps medium \"FB Armada\"" but got "small-caps medium/normal \"FB Armada\""
FAIL e.style['font'] = "small-caps normal xx-large/1.2 serif" should set the property value assert_equals: serialization should be canonical expected "small-caps xx-large / 1.2 serif" but got "normal small-caps xx-large/1.2 serif"
FAIL e.style['font'] = "small-caps normal normal larger/calc(120% + 1.2em) sans-serif" should set the property value assert_equals: serialization should be canonical expected "small-caps larger / calc(120% + 1.2em) sans-serif" but got "normal small-caps normal larger/calc(120% + 1.2em) sans-serif"
FAIL e.style['font'] = "small-caps normal normal normal smaller cursive" should set the property value assert_equals: serialization should be canonical expected "small-caps smaller cursive" but got "normal small-caps normal normal smaller cursive"
FAIL e.style['font'] = "small-caps normal normal italic 10px/normal fantasy" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "small-caps normal normal bolder 20%/1.2 monospace" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "small-caps normal normal semi-expanded calc(30% - 40px)/calc(120% + 1.2em) Menu" should set the property value assert_equals: serialization should be canonical expected "small-caps semi-expanded calc(30% - 40px) / calc(120% + 1.2em) Menu" but got "normal small-caps normal semi-expanded calc(30% - 40px)/calc(120% + 1.2em) Menu"
FAIL e.style['font'] = "small-caps normal oblique xx-small \"FB Armada\"" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "small-caps normal italic normal medium/normal serif" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "small-caps normal oblique lighter xx-large/1.2 sans-serif" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "small-caps normal italic expanded larger/calc(120% + 1.2em) cursive" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "small-caps normal 100 smaller fantasy" should set the property value assert_equals: serialization should be canonical expected "small-caps 100 smaller fantasy" but got "normal small-caps 100 smaller fantasy"
FAIL e.style['font'] = "small-caps normal 900 normal 10px/normal monospace" should set the property value assert_equals: serialization should be canonical expected "small-caps 900 10px monospace" but got "normal small-caps 900 normal 10px/normal monospace"
FAIL e.style['font'] = "small-caps normal bold oblique 20%/1.2 Menu" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "small-caps normal bolder extra-expanded calc(30% - 40px)/calc(120% + 1.2em) \"FB Armada\"" should set the property value assert_equals: serialization should be canonical expected "small-caps bolder extra-expanded calc(30% - 40px) / calc(120% + 1.2em) \"FB Armada\"" but got "normal small-caps bolder extra-expanded calc(30% - 40px)/calc(120% + 1.2em) \"FB Armada\""
FAIL e.style['font'] = "small-caps normal ultra-expanded xx-small serif" should set the property value assert_equals: serialization should be canonical expected "small-caps ultra-expanded xx-small serif" but got "normal small-caps ultra-expanded xx-small serif"
FAIL e.style['font'] = "small-caps normal ultra-condensed normal medium/normal sans-serif" should set the property value assert_equals: serialization should be canonical expected "small-caps ultra-condensed medium sans-serif" but got "normal small-caps normal ultra-condensed medium/normal sans-serif"
FAIL e.style['font'] = "small-caps normal extra-condensed italic xx-large/1.2 cursive" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "small-caps normal condensed lighter larger/calc(120% + 1.2em) fantasy" should set the property value assert_equals: serialization should be canonical expected "small-caps lighter condensed larger / calc(120% + 1.2em) fantasy" but got "normal small-caps lighter condensed larger/calc(120% + 1.2em) fantasy"
PASS e.style['font'] = "small-caps oblique smaller monospace" should set the property value
FAIL e.style['font'] = "small-caps italic normal 10px/normal Menu" should set the property value assert_equals: serialization should be canonical expected "italic small-caps 10px Menu" but got "italic small-caps normal 10px/normal Menu"
FAIL e.style['font'] = "small-caps oblique normal normal 20%/1.2 \"FB Armada\"" should set the property value assert_equals: serialization should be canonical expected "oblique small-caps 20% / 1.2 \"FB Armada\"" but got "oblique small-caps normal normal 20%/1.2 \"FB Armada\""
FAIL e.style['font'] = "small-caps italic normal 100 calc(30% - 40px)/calc(120% + 1.2em) serif" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "small-caps oblique normal semi-condensed xx-small sans-serif" should set the property value assert_equals: serialization should be canonical expected "oblique small-caps semi-condensed xx-small sans-serif" but got "oblique small-caps normal semi-condensed xx-small sans-serif"
FAIL e.style['font'] = "small-caps italic 900 medium/normal cursive" should set the property value assert_equals: serialization should be canonical expected "italic small-caps 900 medium cursive" but got "italic small-caps 900 medium/normal cursive"
FAIL e.style['font'] = "small-caps oblique bold normal xx-large/1.2 fantasy" should set the property value assert_equals: serialization should be canonical expected "oblique small-caps bold xx-large / 1.2 fantasy" but got "oblique small-caps bold normal xx-large/1.2 fantasy"
FAIL e.style['font'] = "small-caps italic bolder semi-expanded larger/calc(120% + 1.2em) monospace" should set the property value assert_equals: serialization should be canonical expected "italic small-caps bolder semi-expanded larger / calc(120% + 1.2em) monospace" but got "italic small-caps bolder semi-expanded larger/calc(120% + 1.2em) monospace"
PASS e.style['font'] = "small-caps oblique expanded smaller Menu" should set the property value
FAIL e.style['font'] = "small-caps italic extra-expanded normal 10px/normal \"FB Armada\"" should set the property value assert_equals: serialization should be canonical expected "italic small-caps extra-expanded 10px \"FB Armada\"" but got "italic small-caps normal extra-expanded 10px/normal \"FB Armada\""
FAIL e.style['font'] = "small-caps oblique ultra-expanded lighter 20%/1.2 serif" should set the property value assert_equals: serialization should be canonical expected "oblique small-caps lighter ultra-expanded 20% / 1.2 serif" but got "oblique small-caps lighter ultra-expanded 20%/1.2 serif"
FAIL e.style['font'] = "small-caps 100 calc(30% - 40px)/calc(120% + 1.2em) sans-serif" should set the property value assert_equals: serialization should be canonical expected "small-caps 100 calc(30% - 40px) / calc(120% + 1.2em) sans-serif" but got "small-caps 100 calc(30% - 40px)/calc(120% + 1.2em) sans-serif"
FAIL e.style['font'] = "small-caps 900 normal xx-small cursive" should set the property value assert_equals: serialization should be canonical expected "small-caps 900 xx-small cursive" but got "normal small-caps 900 xx-small cursive"
FAIL e.style['font'] = "small-caps bold normal normal medium/normal fantasy" should set the property value assert_equals: serialization should be canonical expected "small-caps bold medium fantasy" but got "normal small-caps bold normal medium/normal fantasy"
FAIL e.style['font'] = "small-caps bolder normal italic xx-large/1.2 monospace" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "small-caps lighter normal ultra-condensed larger/calc(120% + 1.2em) Menu" should set the property value assert_equals: serialization should be canonical expected "small-caps lighter ultra-condensed larger / calc(120% + 1.2em) Menu" but got "normal small-caps lighter ultra-condensed larger/calc(120% + 1.2em) Menu"
PASS e.style['font'] = "small-caps 100 oblique smaller \"FB Armada\"" should set the property value
FAIL e.style['font'] = "small-caps 900 italic normal 10px/normal serif" should set the property value assert_equals: serialization should be canonical expected "italic small-caps 900 10px serif" but got "italic small-caps 900 normal 10px/normal serif"
FAIL e.style['font'] = "small-caps bold oblique extra-condensed 20%/1.2 sans-serif" should set the property value assert_equals: serialization should be canonical expected "oblique small-caps bold extra-condensed 20% / 1.2 sans-serif" but got "oblique small-caps bold extra-condensed 20%/1.2 sans-serif"
FAIL e.style['font'] = "small-caps bolder condensed calc(30% - 40px)/calc(120% + 1.2em) cursive" should set the property value assert_equals: serialization should be canonical expected "small-caps bolder condensed calc(30% - 40px) / calc(120% + 1.2em) cursive" but got "small-caps bolder condensed calc(30% - 40px)/calc(120% + 1.2em) cursive"
FAIL e.style['font'] = "small-caps lighter semi-condensed normal xx-small fantasy" should set the property value assert_equals: serialization should be canonical expected "small-caps lighter semi-condensed xx-small fantasy" but got "normal small-caps lighter semi-condensed xx-small fantasy"
FAIL e.style['font'] = "small-caps 100 semi-expanded italic medium/normal monospace" should set the property value assert_equals: serialization should be canonical expected "italic small-caps 100 semi-expanded medium monospace" but got "italic small-caps 100 semi-expanded medium/normal monospace"
FAIL e.style['font'] = "small-caps expanded xx-large/1.2 Menu" should set the property value assert_equals: serialization should be canonical expected "small-caps expanded xx-large / 1.2 Menu" but got "small-caps expanded xx-large/1.2 Menu"
FAIL e.style['font'] = "small-caps extra-expanded normal larger/calc(120% + 1.2em) \"FB Armada\"" should set the property value assert_equals: serialization should be canonical expected "small-caps extra-expanded larger / calc(120% + 1.2em) \"FB Armada\"" but got "normal small-caps extra-expanded larger/calc(120% + 1.2em) \"FB Armada\""
FAIL e.style['font'] = "small-caps ultra-expanded normal normal smaller serif" should set the property value assert_equals: serialization should be canonical expected "small-caps ultra-expanded smaller serif" but got "normal small-caps normal ultra-expanded smaller serif"
FAIL e.style['font'] = "small-caps ultra-condensed normal oblique 10px/normal sans-serif" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "small-caps extra-condensed normal 900 20%/1.2 cursive" should set the property value assert_equals: serialization should be canonical expected "small-caps 900 extra-condensed 20% / 1.2 cursive" but got "normal small-caps 900 extra-condensed 20%/1.2 cursive"
FAIL e.style['font'] = "small-caps condensed italic calc(30% - 40px)/calc(120% + 1.2em) fantasy" should set the property value assert_equals: serialization should be canonical expected "italic small-caps condensed calc(30% - 40px) / calc(120% + 1.2em) fantasy" but got "italic small-caps condensed calc(30% - 40px)/calc(120% + 1.2em) fantasy"
FAIL e.style['font'] = "small-caps semi-condensed oblique normal xx-small monospace" should set the property value assert_equals: serialization should be canonical expected "oblique small-caps semi-condensed xx-small monospace" but got "oblique small-caps normal semi-condensed xx-small monospace"
FAIL e.style['font'] = "small-caps semi-expanded italic bold medium/normal Menu" should set the property value assert_equals: serialization should be canonical expected "italic small-caps bold semi-expanded medium Menu" but got "italic small-caps bold semi-expanded medium/normal Menu"
FAIL e.style['font'] = "small-caps expanded bolder xx-large/1.2 \"FB Armada\"" should set the property value assert_equals: serialization should be canonical expected "small-caps bolder expanded xx-large / 1.2 \"FB Armada\"" but got "small-caps bolder expanded xx-large/1.2 \"FB Armada\""
FAIL e.style['font'] = "small-caps extra-expanded lighter normal larger/calc(120% + 1.2em) serif" should set the property value assert_equals: serialization should be canonical expected "small-caps lighter extra-expanded larger / calc(120% + 1.2em) serif" but got "normal small-caps lighter extra-expanded larger/calc(120% + 1.2em) serif"
PASS e.style['font'] = "small-caps ultra-expanded 100 oblique smaller sans-serif" should set the property value
FAIL e.style['font'] = "900 10px/normal cursive" should set the property value assert_equals: serialization should be canonical expected "900 10px cursive" but got "900 10px/normal cursive"
FAIL e.style['font'] = "bold normal 20%/1.2 fantasy" should set the property value assert_equals: serialization should be canonical expected "bold 20% / 1.2 fantasy" but got "normal bold 20%/1.2 fantasy"
FAIL e.style['font'] = "bolder normal normal calc(30% - 40px)/calc(120% + 1.2em) monospace" should set the property value assert_equals: serialization should be canonical expected "bolder calc(30% - 40px) / calc(120% + 1.2em) monospace" but got "normal normal bolder calc(30% - 40px)/calc(120% + 1.2em) monospace"
FAIL e.style['font'] = "lighter normal normal normal xx-small Menu" should set the property value assert_equals: serialization should be canonical expected "lighter xx-small Menu" but got "normal normal lighter normal xx-small Menu"
FAIL e.style['font'] = "100 normal normal italic medium/normal \"FB Armada\"" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "900 normal normal small-caps xx-large/1.2 serif" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "bold normal normal ultra-condensed larger/calc(120% + 1.2em) sans-serif" should set the property value assert_equals: serialization should be canonical expected "bold ultra-condensed larger / calc(120% + 1.2em) sans-serif" but got "normal normal bold ultra-condensed larger/calc(120% + 1.2em) sans-serif"
FAIL e.style['font'] = "bolder normal oblique smaller cursive" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "lighter normal italic normal 10px/normal fantasy" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "100 normal oblique small-caps 20%/1.2 monospace" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "900 normal italic extra-condensed calc(30% - 40px)/calc(120% + 1.2em) Menu" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "bold normal small-caps xx-small \"FB Armada\"" should set the property value assert_equals: serialization should be canonical expected "small-caps bold xx-small \"FB Armada\"" but got "normal small-caps bold xx-small \"FB Armada\""
FAIL e.style['font'] = "bolder normal small-caps normal medium/normal serif" should set the property value assert_equals: serialization should be canonical expected "small-caps bolder medium serif" but got "normal small-caps bolder normal medium/normal serif"
FAIL e.style['font'] = "lighter normal small-caps oblique xx-large/1.2 sans-serif" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "100 normal small-caps condensed larger/calc(120% + 1.2em) cursive" should set the property value assert_equals: serialization should be canonical expected "small-caps 100 condensed larger / calc(120% + 1.2em) cursive" but got "normal small-caps 100 condensed larger/calc(120% + 1.2em) cursive"
FAIL e.style['font'] = "900 normal semi-condensed smaller fantasy" should set the property value assert_equals: serialization should be canonical expected "900 semi-condensed smaller fantasy" but got "normal 900 semi-condensed smaller fantasy"
FAIL e.style['font'] = "bold normal semi-expanded normal 10px/normal monospace" should set the property value assert_equals: serialization should be canonical expected "bold semi-expanded 10px monospace" but got "normal normal bold semi-expanded 10px/normal monospace"
FAIL e.style['font'] = "bolder normal expanded italic 20%/1.2 Menu" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "lighter normal extra-expanded small-caps calc(30% - 40px)/calc(120% + 1.2em) \"FB Armada\"" should set the property value assert_equals: serialization should be canonical expected "small-caps lighter extra-expanded calc(30% - 40px) / calc(120% + 1.2em) \"FB Armada\"" but got "normal small-caps lighter extra-expanded calc(30% - 40px)/calc(120% + 1.2em) \"FB Armada\""
PASS e.style['font'] = "100 oblique xx-small serif" should set the property value
FAIL e.style['font'] = "900 italic normal medium/normal sans-serif" should set the property value assert_equals: serialization should be canonical expected "italic 900 medium sans-serif" but got "italic normal 900 medium/normal sans-serif"
FAIL e.style['font'] = "bold oblique normal normal xx-large/1.2 cursive" should set the property value assert_equals: serialization should be canonical expected "oblique bold xx-large / 1.2 cursive" but got "oblique normal bold normal xx-large/1.2 cursive"
FAIL e.style['font'] = "bolder italic normal small-caps larger/calc(120% + 1.2em) fantasy" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "lighter oblique normal ultra-expanded smaller monospace" should set the property value assert_equals: serialization should be canonical expected "oblique lighter ultra-expanded smaller monospace" but got "oblique normal lighter ultra-expanded smaller monospace"
FAIL e.style['font'] = "100 italic small-caps 10px/normal Menu" should set the property value assert_equals: serialization should be canonical expected "italic small-caps 100 10px Menu" but got "italic small-caps 100 10px/normal Menu"
FAIL e.style['font'] = "900 oblique small-caps normal 20%/1.2 \"FB Armada\"" should set the property value assert_equals: serialization should be canonical expected "oblique small-caps 900 20% / 1.2 \"FB Armada\"" but got "oblique small-caps 900 normal 20%/1.2 \"FB Armada\""
FAIL e.style['font'] = "bold italic small-caps ultra-condensed calc(30% - 40px)/calc(120% + 1.2em) serif" should set the property value assert_equals: serialization should be canonical expected "italic small-caps bold ultra-condensed calc(30% - 40px) / calc(120% + 1.2em) serif" but got "italic small-caps bold ultra-condensed calc(30% - 40px)/calc(120% + 1.2em) serif"
PASS e.style['font'] = "bolder oblique extra-condensed xx-small sans-serif" should set the property value
FAIL e.style['font'] = "lighter italic condensed normal medium/normal cursive" should set the property value assert_equals: serialization should be canonical expected "italic lighter condensed medium cursive" but got "italic normal lighter condensed medium/normal cursive"
FAIL e.style['font'] = "100 oblique semi-condensed small-caps xx-large/1.2 fantasy" should set the property value assert_equals: serialization should be canonical expected "oblique small-caps 100 semi-condensed xx-large / 1.2 fantasy" but got "oblique small-caps 100 semi-condensed xx-large/1.2 fantasy"
FAIL e.style['font'] = "900 small-caps larger/calc(120% + 1.2em) monospace" should set the property value assert_equals: serialization should be canonical expected "small-caps 900 larger / calc(120% + 1.2em) monospace" but got "small-caps 900 larger/calc(120% + 1.2em) monospace"
FAIL e.style['font'] = "bold small-caps normal smaller Menu" should set the property value assert_equals: serialization should be canonical expected "small-caps bold smaller Menu" but got "normal small-caps bold smaller Menu"
FAIL e.style['font'] = "bolder small-caps normal normal 10px/normal \"FB Armada\"" should set the property value assert_equals: serialization should be canonical expected "small-caps bolder 10px \"FB Armada\"" but got "normal small-caps bolder normal 10px/normal \"FB Armada\""
FAIL e.style['font'] = "lighter small-caps normal italic 20%/1.2 serif" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "100 small-caps normal semi-expanded calc(30% - 40px)/calc(120% + 1.2em) sans-serif" should set the property value assert_equals: serialization should be canonical expected "small-caps 100 semi-expanded calc(30% - 40px) / calc(120% + 1.2em) sans-serif" but got "normal small-caps 100 semi-expanded calc(30% - 40px)/calc(120% + 1.2em) sans-serif"
PASS e.style['font'] = "900 small-caps oblique xx-small cursive" should set the property value
FAIL e.style['font'] = "bold small-caps italic normal medium/normal fantasy" should set the property value assert_equals: serialization should be canonical expected "italic small-caps bold medium fantasy" but got "italic small-caps bold normal medium/normal fantasy"
FAIL e.style['font'] = "bolder small-caps oblique expanded xx-large/1.2 monospace" should set the property value assert_equals: serialization should be canonical expected "oblique small-caps bolder expanded xx-large / 1.2 monospace" but got "oblique small-caps bolder expanded xx-large/1.2 monospace"
FAIL e.style['font'] = "lighter small-caps extra-expanded larger/calc(120% + 1.2em) Menu" should set the property value assert_equals: serialization should be canonical expected "small-caps lighter extra-expanded larger / calc(120% + 1.2em) Menu" but got "small-caps lighter extra-expanded larger/calc(120% + 1.2em) Menu"
FAIL e.style['font'] = "100 small-caps ultra-expanded normal smaller \"FB Armada\"" should set the property value assert_equals: serialization should be canonical expected "small-caps 100 ultra-expanded smaller \"FB Armada\"" but got "normal small-caps 100 ultra-expanded smaller \"FB Armada\""
FAIL e.style['font'] = "900 small-caps ultra-condensed italic 10px/normal serif" should set the property value assert_equals: serialization should be canonical expected "italic small-caps 900 ultra-condensed 10px serif" but got "italic small-caps 900 ultra-condensed 10px/normal serif"
FAIL e.style['font'] = "bold extra-condensed 20%/1.2 sans-serif" should set the property value assert_equals: serialization should be canonical expected "bold extra-condensed 20% / 1.2 sans-serif" but got "bold extra-condensed 20%/1.2 sans-serif"
FAIL e.style['font'] = "bolder condensed normal calc(30% - 40px)/calc(120% + 1.2em) cursive" should set the property value assert_equals: serialization should be canonical expected "bolder condensed calc(30% - 40px) / calc(120% + 1.2em) cursive" but got "normal bolder condensed calc(30% - 40px)/calc(120% + 1.2em) cursive"
FAIL e.style['font'] = "lighter semi-condensed normal normal xx-small fantasy" should set the property value assert_equals: serialization should be canonical expected "lighter semi-condensed xx-small fantasy" but got "normal normal lighter semi-condensed xx-small fantasy"
FAIL e.style['font'] = "100 semi-expanded normal oblique medium/normal monospace" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "900 expanded normal small-caps xx-large/1.2 Menu" should set the property value assert_equals: serialization should be canonical expected "small-caps 900 expanded xx-large / 1.2 Menu" but got "normal small-caps 900 expanded xx-large/1.2 Menu"
FAIL e.style['font'] = "bold extra-expanded italic larger/calc(120% + 1.2em) \"FB Armada\"" should set the property value assert_equals: serialization should be canonical expected "italic bold extra-expanded larger / calc(120% + 1.2em) \"FB Armada\"" but got "italic bold extra-expanded larger/calc(120% + 1.2em) \"FB Armada\""
FAIL e.style['font'] = "bolder ultra-expanded oblique normal smaller serif" should set the property value assert_equals: serialization should be canonical expected "oblique bolder ultra-expanded smaller serif" but got "oblique normal bolder ultra-expanded smaller serif"
FAIL e.style['font'] = "lighter ultra-condensed italic small-caps 10px/normal sans-serif" should set the property value assert_equals: serialization should be canonical expected "italic small-caps lighter ultra-condensed 10px sans-serif" but got "italic small-caps lighter ultra-condensed 10px/normal sans-serif"
FAIL e.style['font'] = "100 extra-condensed small-caps 20%/1.2 cursive" should set the property value assert_equals: serialization should be canonical expected "small-caps 100 extra-condensed 20% / 1.2 cursive" but got "small-caps 100 extra-condensed 20%/1.2 cursive"
FAIL e.style['font'] = "900 condensed small-caps normal calc(30% - 40px)/calc(120% + 1.2em) fantasy" should set the property value assert_equals: serialization should be canonical expected "small-caps 900 condensed calc(30% - 40px) / calc(120% + 1.2em) fantasy" but got "normal small-caps 900 condensed calc(30% - 40px)/calc(120% + 1.2em) fantasy"
PASS e.style['font'] = "bold semi-condensed small-caps oblique xx-small monospace" should set the property value
FAIL e.style['font'] = "semi-expanded medium/normal Menu" should set the property value assert_equals: serialization should be canonical expected "semi-expanded medium Menu" but got "semi-expanded medium/normal Menu"
FAIL e.style['font'] = "expanded normal xx-large/1.2 \"FB Armada\"" should set the property value assert_equals: serialization should be canonical expected "expanded xx-large / 1.2 \"FB Armada\"" but got "normal expanded xx-large/1.2 \"FB Armada\""
FAIL e.style['font'] = "extra-expanded normal normal larger/calc(120% + 1.2em) serif" should set the property value assert_equals: serialization should be canonical expected "extra-expanded larger / calc(120% + 1.2em) serif" but got "normal normal extra-expanded larger/calc(120% + 1.2em) serif"
FAIL e.style['font'] = "ultra-expanded normal normal normal smaller sans-serif" should set the property value assert_equals: serialization should be canonical expected "ultra-expanded smaller sans-serif" but got "normal normal normal ultra-expanded smaller sans-serif"
FAIL e.style['font'] = "ultra-condensed normal normal italic 10px/normal cursive" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "extra-condensed normal normal small-caps 20%/1.2 fantasy" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "condensed normal normal bolder calc(30% - 40px)/calc(120% + 1.2em) monospace" should set the property value assert_equals: serialization should be canonical expected "bolder condensed calc(30% - 40px) / calc(120% + 1.2em) monospace" but got "normal normal bolder condensed calc(30% - 40px)/calc(120% + 1.2em) monospace"
FAIL e.style['font'] = "semi-condensed normal oblique xx-small Menu" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "semi-expanded normal italic normal medium/normal \"FB Armada\"" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "expanded normal oblique small-caps xx-large/1.2 serif" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "extra-expanded normal italic lighter larger/calc(120% + 1.2em) sans-serif" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "ultra-expanded normal small-caps smaller cursive" should set the property value assert_equals: serialization should be canonical expected "small-caps ultra-expanded smaller cursive" but got "normal small-caps ultra-expanded smaller cursive"
FAIL e.style['font'] = "ultra-condensed normal small-caps normal 10px/normal fantasy" should set the property value assert_equals: serialization should be canonical expected "small-caps ultra-condensed 10px fantasy" but got "normal small-caps normal ultra-condensed 10px/normal fantasy"
FAIL e.style['font'] = "extra-condensed normal small-caps oblique 20%/1.2 monospace" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "condensed normal small-caps 100 calc(30% - 40px)/calc(120% + 1.2em) Menu" should set the property value assert_equals: serialization should be canonical expected "small-caps 100 condensed calc(30% - 40px) / calc(120% + 1.2em) Menu" but got "normal small-caps 100 condensed calc(30% - 40px)/calc(120% + 1.2em) Menu"
FAIL e.style['font'] = "semi-condensed normal 900 xx-small \"FB Armada\"" should set the property value assert_equals: serialization should be canonical expected "900 semi-condensed xx-small \"FB Armada\"" but got "normal 900 semi-condensed xx-small \"FB Armada\""
FAIL e.style['font'] = "semi-expanded normal bold normal medium/normal serif" should set the property value assert_equals: serialization should be canonical expected "bold semi-expanded medium serif" but got "normal normal bold semi-expanded medium/normal serif"
FAIL e.style['font'] = "expanded normal bolder italic xx-large/1.2 sans-serif" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "extra-expanded normal lighter small-caps larger/calc(120% + 1.2em) cursive" should set the property value assert_equals: serialization should be canonical expected "small-caps lighter extra-expanded larger / calc(120% + 1.2em) cursive" but got "normal small-caps lighter extra-expanded larger/calc(120% + 1.2em) cursive"
PASS e.style['font'] = "ultra-expanded oblique smaller fantasy" should set the property value
FAIL e.style['font'] = "ultra-condensed italic normal 10px/normal monospace" should set the property value assert_equals: serialization should be canonical expected "italic ultra-condensed 10px monospace" but got "italic normal ultra-condensed 10px/normal monospace"
FAIL e.style['font'] = "extra-condensed oblique normal normal 20%/1.2 Menu" should set the property value assert_equals: serialization should be canonical expected "oblique extra-condensed 20% / 1.2 Menu" but got "oblique normal normal extra-condensed 20%/1.2 Menu"
FAIL e.style['font'] = "condensed italic normal small-caps calc(30% - 40px)/calc(120% + 1.2em) \"FB Armada\"" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "semi-condensed oblique normal 100 xx-small serif" should set the property value assert_equals: serialization should be canonical expected "oblique 100 semi-condensed xx-small serif" but got "oblique normal 100 semi-condensed xx-small serif"
FAIL e.style['font'] = "semi-expanded italic small-caps medium/normal sans-serif" should set the property value assert_equals: serialization should be canonical expected "italic small-caps semi-expanded medium sans-serif" but got "italic small-caps semi-expanded medium/normal sans-serif"
FAIL e.style['font'] = "expanded oblique small-caps normal xx-large/1.2 cursive" should set the property value assert_equals: serialization should be canonical expected "oblique small-caps expanded xx-large / 1.2 cursive" but got "oblique small-caps normal expanded xx-large/1.2 cursive"
FAIL e.style['font'] = "extra-expanded italic small-caps 900 larger/calc(120% + 1.2em) fantasy" should set the property value assert_equals: serialization should be canonical expected "italic small-caps 900 extra-expanded larger / calc(120% + 1.2em) fantasy" but got "italic small-caps 900 extra-expanded larger/calc(120% + 1.2em) fantasy"
PASS e.style['font'] = "ultra-expanded oblique bold smaller monospace" should set the property value
FAIL e.style['font'] = "ultra-condensed italic bolder normal 10px/normal Menu" should set the property value assert_equals: serialization should be canonical expected "italic bolder ultra-condensed 10px Menu" but got "italic normal bolder ultra-condensed 10px/normal Menu"
FAIL e.style['font'] = "extra-condensed oblique lighter small-caps 20%/1.2 \"FB Armada\"" should set the property value assert_equals: serialization should be canonical expected "oblique small-caps lighter extra-condensed 20% / 1.2 \"FB Armada\"" but got "oblique small-caps lighter extra-condensed 20%/1.2 \"FB Armada\""
FAIL e.style['font'] = "condensed small-caps calc(30% - 40px)/calc(120% + 1.2em) serif" should set the property value assert_equals: serialization should be canonical expected "small-caps condensed calc(30% - 40px) / calc(120% + 1.2em) serif" but got "small-caps condensed calc(30% - 40px)/calc(120% + 1.2em) serif"
FAIL e.style['font'] = "semi-condensed small-caps normal xx-small sans-serif" should set the property value assert_equals: serialization should be canonical expected "small-caps semi-condensed xx-small sans-serif" but got "normal small-caps semi-condensed xx-small sans-serif"
FAIL e.style['font'] = "semi-expanded small-caps normal normal medium/normal cursive" should set the property value assert_equals: serialization should be canonical expected "small-caps semi-expanded medium cursive" but got "normal small-caps normal semi-expanded medium/normal cursive"
FAIL e.style['font'] = "expanded small-caps normal italic xx-large/1.2 fantasy" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "extra-expanded small-caps normal 100 larger/calc(120% + 1.2em) monospace" should set the property value assert_equals: serialization should be canonical expected "small-caps 100 extra-expanded larger / calc(120% + 1.2em) monospace" but got "normal small-caps 100 extra-expanded larger/calc(120% + 1.2em) monospace"
PASS e.style['font'] = "ultra-expanded small-caps oblique smaller Menu" should set the property value
FAIL e.style['font'] = "ultra-condensed small-caps italic normal 10px/normal \"FB Armada\"" should set the property value assert_equals: serialization should be canonical expected "italic small-caps ultra-condensed 10px \"FB Armada\"" but got "italic small-caps normal ultra-condensed 10px/normal \"FB Armada\""
FAIL e.style['font'] = "extra-condensed small-caps oblique 900 20%/1.2 serif" should set the property value assert_equals: serialization should be canonical expected "oblique small-caps 900 extra-condensed 20% / 1.2 serif" but got "oblique small-caps 900 extra-condensed 20%/1.2 serif"
FAIL e.style['font'] = "condensed small-caps bold calc(30% - 40px)/calc(120% + 1.2em) sans-serif" should set the property value assert_equals: serialization should be canonical expected "small-caps bold condensed calc(30% - 40px) / calc(120% + 1.2em) sans-serif" but got "small-caps bold condensed calc(30% - 40px)/calc(120% + 1.2em) sans-serif"
FAIL e.style['font'] = "semi-condensed small-caps bolder normal xx-small cursive" should set the property value assert_equals: serialization should be canonical expected "small-caps bolder semi-condensed xx-small cursive" but got "normal small-caps bolder semi-condensed xx-small cursive"
FAIL e.style['font'] = "semi-expanded small-caps lighter italic medium/normal fantasy" should set the property value assert_equals: serialization should be canonical expected "italic small-caps lighter semi-expanded medium fantasy" but got "italic small-caps lighter semi-expanded medium/normal fantasy"
FAIL e.style['font'] = "expanded 100 xx-large/1.2 monospace" should set the property value assert_equals: serialization should be canonical expected "100 expanded xx-large / 1.2 monospace" but got "100 expanded xx-large/1.2 monospace"
FAIL e.style['font'] = "extra-expanded 900 normal larger/calc(120% + 1.2em) Menu" should set the property value assert_equals: serialization should be canonical expected "900 extra-expanded larger / calc(120% + 1.2em) Menu" but got "normal 900 extra-expanded larger/calc(120% + 1.2em) Menu"
FAIL e.style['font'] = "ultra-expanded bold normal normal smaller \"FB Armada\"" should set the property value assert_equals: serialization should be canonical expected "bold ultra-expanded smaller \"FB Armada\"" but got "normal normal bold ultra-expanded smaller \"FB Armada\""
FAIL e.style['font'] = "ultra-condensed bolder normal oblique 10px/normal serif" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['font'] = "extra-condensed lighter normal small-caps 20%/1.2 sans-serif" should set the property value assert_equals: serialization should be canonical expected "small-caps lighter extra-condensed 20% / 1.2 sans-serif" but got "normal small-caps lighter extra-condensed 20%/1.2 sans-serif"
FAIL e.style['font'] = "condensed 100 italic calc(30% - 40px)/calc(120% + 1.2em) cursive" should set the property value assert_equals: serialization should be canonical expected "italic 100 condensed calc(30% - 40px) / calc(120% + 1.2em) cursive" but got "italic 100 condensed calc(30% - 40px)/calc(120% + 1.2em) cursive"
FAIL e.style['font'] = "semi-condensed 900 oblique normal xx-small fantasy" should set the property value assert_equals: serialization should be canonical expected "oblique 900 semi-condensed xx-small fantasy" but got "oblique normal 900 semi-condensed xx-small fantasy"
FAIL e.style['font'] = "semi-expanded bold italic small-caps medium/normal monospace" should set the property value assert_equals: serialization should be canonical expected "italic small-caps bold semi-expanded medium monospace" but got "italic small-caps bold semi-expanded medium/normal monospace"
FAIL e.style['font'] = "expanded bolder small-caps xx-large/1.2 Menu" should set the property value assert_equals: serialization should be canonical expected "small-caps bolder expanded xx-large / 1.2 Menu" but got "small-caps bolder expanded xx-large/1.2 Menu"
FAIL e.style['font'] = "extra-expanded lighter small-caps normal larger/calc(120% + 1.2em) \"FB Armada\"" should set the property value assert_equals: serialization should be canonical expected "small-caps lighter extra-expanded larger / calc(120% + 1.2em) \"FB Armada\"" but got "normal small-caps lighter extra-expanded larger/calc(120% + 1.2em) \"FB Armada\""
PASS e.style['font'] = "ultra-expanded 100 small-caps oblique smaller serif" should set the property value
